pyproject.toml
[tool.poetry]
name = "battleship"
version = "0.1.0"
description = "a framework for battleship game"
authors = ["mike239x"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = ">=3.5"
pyzmq = "^18.0"
numpy = "^1.16"
scipy = "^1.2"
scikit-image = "^0.15.0"
PyQt5 = "^5.12"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
